Eliminate MSVC warnings in cckddasd.c cckddasd.c(1580) : error C2036: 'void *' : unknown size cckddasd.c(1723) : error C2036: 'void *' : unknown size cckddasd.c(4203) : warning C4244: 'function' : conversion from '__int64' to 'unsigned int', possible loss of data
git-svn-id: file:///home/jj/hercules.svn/trunk@2610 956126f8-22a0-4046-8f4a-272fa8102e63

+++ cckddasd.c
@@ -1577,7 +1577,7 @@ TID             tid;                    /* Writer thread id          */
             signal_condition (&cckdblk.wrcond);
         else if (cckdblk.wrs < cckdblk.wrmax)
         {
-            rc = create_thread (&tid, NULL, cckd_writer, (void *)(long)cckdblk.wrs + 1);
+            rc = create_thread (&tid, NULL, cckd_writer, (void *)((long)cckdblk.wrs + 1));
             if (rc == 0) cckdblk.wrs++;
         }
     }
@@ -1720,7 +1720,7 @@ BYTE            buf2[65536];            /* Compress buffer           */
                 signal_condition (&cckdblk.wrcond);
             else if (cckdblk.wrs < cckdblk.wrmax)
             {
-                rc = create_thread (&tid, NULL, cckd_writer, (void *)(long)cckdblk.wrs + 1);
+                rc = create_thread (&tid, NULL, cckd_writer, (void *)((long)cckdblk.wrs + 1));
                 if (rc == 0) cckdblk.wrs++;
             }
         }
@@ -4200,7 +4200,7 @@ int             gctab[5]= {             /* default gcol parameters   */
             release_lock (&cckd->iolock);
 
             /* Call the garbage collector */
-            cckd_gc_percolate (dev, size);
+            cckd_gc_percolate (dev, (unsigned int)size);
 
             /* Schedule any updated tracks to be written */
             obtain_lock (&cckd->iolock);
